huge refactoring
