+++ src/redux/sagas/submissions.saga.js
@@ -6,7 +6,6 @@ function* submissionsSaga() {
   yield takeLatest('FETCH_SUBMISSIONS', fetchSubmissions);
   yield takeLatest('FETCH_DESCRIPTION', fetchDescription);
   yield takeLatest('UPDATE_FORM_STATUS', updateFormStatus);
-  yield takeLatest('UPDATE_LIKES', updateLikes);
   yield takeLatest('FETCH_ALL_CATEGORIES', fetchAllCategories);
 }
 
@@ -83,21 +82,6 @@ function* updateFormStatus(action) {
   }
 }
 
-function* updateLikes(action) {
-  console.log('action.payload of updateLikes');
-  try {
-    yield axios({
-      method: 'PUT',
-      url: `/api/submissions/likes/${action.payload}`,
-    });
 
-    yield put({
-      method: 'PUT',
-      type: 'FETCH_SUBMISSIONS',
-    });
-  } catch (err) {
-    console.error('ERROR in updateLikes saga:', err);
-  }
-}
 
 export default submissionsSaga;
